Global refactor backend

+++ backend/src/modules/job-vacancy/controller.js
@@ -6,7 +6,7 @@ const toGetJobVacancyList = (req, res) => {
     try {
         vacancyDataBase.getJobVacancyList()
             .then(async vacancyList => {
-                workers.postWorkerMessage("ImageConverterWorker", {
+                workers.postWorkerMessage("FileWorker", {
                     method: "check",
                     data: vacancyList
                 }, (result) => {
@@ -19,8 +19,15 @@ const toGetJobVacancyList = (req, res) => {
 };
 const toGetJobVacancy = async (req, res) => {
     try {
-        const data = await vacancyDataBase.getJobVacancy(req.params);
-        res.status(200).json(data);
+        vacancyDataBase.getJobVacancy(req.params.id)
+            .then(async vacancy =>{
+                workers.postWorkerMessage("FileWorker", {
+                    method: "check",
+                    data: vacancy
+                }, (result) => {
+                    res.status(200).send(result)
+                })
+            })
     } catch (e) {
         res.status(404).send();
     }
@@ -29,7 +36,7 @@ const toPostJobVacancy = (req, res) => {
     try {
         const imagePath = process.cwd()+process.env.FILES_STATIC_IMAGES_VACANCY_FOLDER;
         vacancyDataBase.postJobVacancy(req.query).then(id => {
-            workers.postWorkerMessage("ImageConverterWorker", {
+            workers.postWorkerMessage("FileWorker", {
                 method: "convert",
                 data: {
                     pathImage: req.files.vacancy[0].path,
@@ -42,18 +49,37 @@ const toPostJobVacancy = (req, res) => {
         })
     } catch (e) {
         res.status(404).send();
-        console.log(e);
+        console.error(e);
     }
 };
+const toUpdateJobVacancy = async (req,res)=>{
+    vacancyDataBase.updateJobVacancy(req.params,req.query)
+        .then(id => {
+            const imagePath = process.cwd()+process.env.FILES_STATIC_IMAGES_VACANCY_FOLDER;
+            workers.postWorkerMessage("FileWorker", {
+                method: "update",
+                data: {
+                    pathImage: req.files.vacancy[0].path,
+                    toFolder: imagePath,
+                    id,
+                    table: 'vacancy_file'
+                },
+            });
+            res.status(200).send();
+        })
+}
 const toDeleteJobVacancy = async (req, res) => {
     try {
         vacancyDataBase.deleteJobVacancy(req.params)
             .then(vacancy => {
-                workers.postWorkerMessage("ImageConverterWorker", {
-                    method: "delete",
-                    data: {imagePath: vacancy.path},
-                })
-                res.status(200).send();
+                if(!!vacancy){
+                    workers.postWorkerMessage("FileWorker", {
+                        method: "delete",
+                        data: {imagePath: vacancy.path},
+                    })
+                    res.status(200).send();
+                }
+                res.status(404).send();
             })
     } catch (e) {
         res.status(404).send();
@@ -64,5 +90,6 @@ module.exports = {
     toGetJobVacancyList,
     toGetJobVacancy,
     toPostJobVacancy,
+    toUpdateJobVacancy,
     toDeleteJobVacancy,
 };
